:lipstick: feat: create MenuStyle
